moving routing + badge api to sub package
